+++ api/src/controllers/referent.js
@@ -425,7 +425,12 @@ router.put("/young/:id/change-cohort", passport.authenticate("referent", { sessi
       young.set({ meetingPointId: undefined });
     }
 
-    young.set({ statusPhase1: YOUNG_STATUS_PHASE1.WAITING_AFFECTATION, cohort, cohortChangeReason });
+    // si le volontaire change pour la première fois de cohorte, on stocke sa cohorte d'origine
+    if (!young.originalCohort) {
+      young.set({ originalCohort: young.cohort });
+    }
+
+    young.set({ statusPhase1: YOUNG_STATUS_PHASE1.WAITING_AFFECTATION, cohort, cohortChangeReason, cohesionStayPresence: undefined, cohesionStayMedicalFileReceived: undefined });
     await young.save({ fromUser: req.user });
 
     // if they had a session, we check if we need to update the places taken / left