+++ api/src/models/young.js
@@ -89,6 +89,13 @@ const Schema = new mongoose.Schema({
       description: "Cohorte",
     },
   },
+  originalCohort: {
+    type: String,
+    enum: ["Juillet 2022", "Juin 2022", "Février 2022", "2022", "2021", "2020", "2019"],
+    documentation: {
+      description: "Cohorte d'origine du volontaire, dans le cas ou il a changé de cohorte après sa validation",
+    },
+  },
   cohortChangeReason: {
     type: String,
     documentation: {
@@ -300,7 +307,6 @@ const Schema = new mongoose.Schema({
   cohesionStayPresence: {
     type: String,
     enum: ["true", "false", ""],
-    default: "",
     documentation: {
       description: "Le volontaire était présent lors du séjour de cohésion",
     },
@@ -308,7 +314,6 @@ const Schema = new mongoose.Schema({
   cohesionStayMedicalFileReceived: {
     type: String,
     enum: ["true", "false", ""],
-    default: "",
     documentation: {
       description: "La fiche sanitaire a été reçu par le SNU",
     },