+++ support/src/pages/admin/base-de-connaissance/index.js
@@ -1,3 +1,0 @@
-import KnowledgeBase from "./[slug]";
-
-export default KnowledgeBase;